[Chore/#8] Refactor code
